* commit '60ac9b824f38409519a22b6cf2926f8cf234995f':
  better cleanup
  updated notebooks to reflect new keyword (add_file is now associate for example)
  flake8ed
  ok associate and deassociate seem to work even in sync mode
  updated search doc strings to reflect special meaning of 'file' key
  Associate file and sync seem to work
  pushing so i can work on it offline
  insert_many and delete_many have actually been merged into insert and delete
  insert_many and delete_many seem to have disappeared from sina
  looks like we gt sync going. Need to check add_file still

+++ tests/koshbase.py
@@ -16,23 +16,27 @@ for name in ["sina.datastores.sql", "sina.model", "sina.utils"]:
 
 
 class KoshTest(unittest.TestCase):
-    def init_db(self, engine=None):
+    def init_db(self, engine=None, db_uri=None):
         if engine is None:
             engine = os.environ.get("KOSH_ENGINE", "sina")
         if engine == "sina":
             # Make sure local file is new sql file
-            kosh_test_sql_file = "kosh_test_{}.sql".format(uuid.uuid1().hex)
-            if os.path.exists(kosh_test_sql_file):
+            if db_uri is None:
+                kosh_test_sql_file = "kosh_test_{}.sql".format(uuid.uuid1().hex)
+            else:
+                kosh_test_sql_file = db_uri
+            if db_uri is None and os.path.exists(kosh_test_sql_file):
                 os.remove(kosh_test_sql_file)
-            kosh.utils.create_new_db(kosh_test_sql_file[:-4])
+            if db_uri is None:
+                kosh.utils.create_new_db(kosh_test_sql_file[:-4])
             return kosh_test_sql_file
 
-    def connect(self, engine=None):
+    def connect(self, engine=None, db_uri=None, sync=True):
         if engine is None:
             engine = os.environ.get("KOSH_ENGINE", "sina")
-        kosh_db = self.init_db(engine)
+        kosh_db = self.init_db(engine, db_uri)
         if engine == "sina":
             # os.getlogin does not work on my WSL
             store = KoshStore(engine="sina", username=os.environ["USER"], db='sql',
-                      db_uri=kosh_db)
+                      db_uri=kosh_db, sync=sync)
         return store, kosh_db
